requirements.txt
dirsync==2.2.2
envparse==0.2.0
google-api-python-client==1.6.4
pydrive==1.3.1
python-pushover==0.3
